Refactor Vector2 serialization and deserialization
This commit refactors the serialization and deserialization of the Vector2 struct in the code. It introduces a custom wrapper called Vector2Wrapper to support serde and Visit. The From and Into implementations are added to convert between Vector2 and Vector2Wrapper. The Visit trait is implemented for Vector2Wrapper to handle serialization and deserialization. This change improves the handling of Vector2 in the codebase.